add pop2_render

+++ src/QuickDrawFormats.cc
@@ -22,7 +22,7 @@ using namespace std;
 
 
 
-Color8::Color8(uint16_t r, uint16_t g, uint16_t b) : r(r), g(g), b(b) { }
+Color8::Color8(uint8_t r, uint8_t g, uint8_t b) : r(r), g(g), b(b) { }
